+++ firebase-local/mockData/auth_export/accounts.json
@@ -1,1 +1,86 @@
-{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"TpyNkRxiWfPwunyYEO9fIXlQ26bc","createdAt":"1692721080976","lastLoginAt":"1692721080976","displayName":"Olive Olive","providerUserInfo":[{"providerId":"google.com","rawId":"0770208353068773762792416234417927034661","federatedId":"0770208353068773762792416234417927034661","displayName":"Olive Olive","email":"olive.olive.184@example.com","screenName":"olive_olive"}],"validSince":"1726602821","email":"olive.olive.184@example.com","emailVerified":true,"disabled":false},{"localId":"klTTYGLJGgW0ywpOPYTne2SNYHwA","createdAt":"1692721087300","lastLoginAt":"1692721087300","displayName":"Panda Grass","providerUserInfo":[{"providerId":"google.com","rawId":"6159546765181829175240586069791907892733","federatedId":"6159546765181829175240586069791907892733","displayName":"Panda Grass","email":"panda.grass.494@example.com","screenName":"grass_panda"}],"validSince":"1726602821","email":"panda.grass.494@example.com","emailVerified":true,"disabled":false},{"localId":"orgMemberId","createdAt":"1692117110676","lastLoginAt":"1726602826393","displayName":"Otter Peach","providerUserInfo":[{"providerId":"google.com","rawId":"1869860255802720962031731479365728731411","federatedId":"1869860255802720962031731479365728731411","displayName":"Otter Peach","email":"otter.peach.926@example.com","screenName":"peach_otter"}],"validSince":"1726602821","email":"otter.peach.926@example.com","emailVerified":true,"disabled":false,"lastRefreshAt":"2024-09-17T19:53:46.393Z"},{"localId":"proUserId","createdAt":"1692721093551","lastLoginAt":"1692721093551","displayName":"Olive Grass","providerUserInfo":[{"providerId":"google.com","rawId":"9979981735152850310427424154502370372440","federatedId":"9979981735152850310427424154502370372440","displayName":"Olive Grass","email":"olive.grass.310@example.com","screenName":"grass_olive"}],"validSince":"1726602821","email":"olive.grass.310@example.com","emailVerified":true,"disabled":false}]}
+{
+  "kind": "identitytoolkit#DownloadAccountResponse",
+  "users": [
+    {
+      "localId": "TpyNkRxiWfPwunyYEO9fIXlQ26bc",
+      "createdAt": "1692721080976",
+      "lastLoginAt": "1692721080976",
+      "displayName": "Olive Olive",
+      "providerUserInfo": [
+        {
+          "providerId": "google.com",
+          "rawId": "0770208353068773762792416234417927034661",
+          "federatedId": "0770208353068773762792416234417927034661",
+          "displayName": "Olive Olive",
+          "email": "olive.olive.184@example.com",
+          "screenName": "olive_olive"
+        }
+      ],
+      "validSince": "1726602821",
+      "email": "olive.olive.184@example.com",
+      "emailVerified": true,
+      "disabled": false
+    },
+    {
+      "localId": "klTTYGLJGgW0ywpOPYTne2SNYHwA",
+      "createdAt": "1692721087300",
+      "lastLoginAt": "1692721087300",
+      "displayName": "Panda Grass",
+      "providerUserInfo": [
+        {
+          "providerId": "google.com",
+          "rawId": "6159546765181829175240586069791907892733",
+          "federatedId": "6159546765181829175240586069791907892733",
+          "displayName": "Panda Grass",
+          "email": "panda.grass.494@example.com",
+          "screenName": "grass_panda"
+        }
+      ],
+      "validSince": "1726602821",
+      "email": "panda.grass.494@example.com",
+      "emailVerified": true,
+      "disabled": false
+    },
+    {
+      "localId": "god",
+      "createdAt": "1692117110676",
+      "lastLoginAt": "1726602826393",
+      "displayName": "Otter Peach (God)",
+      "providerUserInfo": [
+        {
+          "providerId": "google.com",
+          "rawId": "1869860255802720962031731479365728731411",
+          "federatedId": "1869860255802720962031731479365728731411",
+          "displayName": "Otter Peach (God)",
+          "email": "otter.peach.926@example.com",
+          "screenName": "peach_otter"
+        }
+      ],
+      "validSince": "1726602821",
+      "email": "otter.peach.926@example.com",
+      "emailVerified": true,
+      "disabled": false,
+      "lastRefreshAt": "2024-09-17T19:53:46.393Z"
+    },
+    {
+      "localId": "proUserId",
+      "createdAt": "1692721093551",
+      "lastLoginAt": "1692721093551",
+      "displayName": "Pro User",
+      "providerUserInfo": [
+        {
+          "providerId": "google.com",
+          "rawId": "9979981735152850310427424154502370372440",
+          "federatedId": "9979981735152850310427424154502370372440",
+          "displayName": "Pro User",
+          "email": "olive.grass.310@example.com",
+          "screenName": "grass_olive"
+        }
+      ],
+      "validSince": "1726602821",
+      "email": "olive.grass.310@example.com",
+      "emailVerified": true,
+      "disabled": false
+    }
+  ]
+}
