+++ src/routes/api/decks/user/+server.ts
@@ -1,4 +1,4 @@
-import { authenticate, getOrganizationDecks, getOrganizationInfo, getUserOrganizations, readPath } from '$lib/server/firebaseUtils';
+import { authenticate, getOrganizationDecks, getOrganizationPublicInfo, getUserOrganizationIds, readPath } from '$lib/server/firebaseUtils';
 import { json } from '@sveltejs/kit';
 import type { RequestHandler } from './$types';
 
@@ -12,10 +12,10 @@ export const GET = (async (event) => {
     const deckArray = Object.entries(userDecks).map(([key, val]) => val);
     return json(deckArray);
   } else if (orgIdParam) {
-    const organizationIds = await getUserOrganizations(uid);
+    const organizationIds = await getUserOrganizationIds(uid);
     if (!organizationIds.includes(orgIdParam)) return new Response('User is not a member of this organization', { status: 403 });
     const orgDecks = (await getOrganizationDecks(orgIdParam)) ?? {};
-    const organizationInfo = await getOrganizationInfo(orgIdParam);
+    const organizationInfo = await getOrganizationPublicInfo(orgIdParam);
     const deckArray = Object.values(orgDecks).map(({ deck }) => ({
       ...deck,
       orgSource: {
@@ -28,12 +28,12 @@ export const GET = (async (event) => {
     const userDecks = (await readPath<Database.Decks.User>(`/decks/user/${uid}`)) || {};
     const userDeckArray = Object.entries(userDecks).map(([key, val]) => val);
     const positionOffset = Math.max(...userDeckArray.map(({ position }) => position), 0) + 2; // +2 because legacy position can be -1, and we want the new position to be strictly greater TODO: refactor this once we get rid of the legacy magic numbers in positioning
-    const organizationIds = await getUserOrganizations(uid);
+    const organizationIds = await getUserOrganizationIds(uid);
     const organizationDeckArray = (
       await Promise.all(
         organizationIds.map(async (orgId) => {
           const orgDecks = (await getOrganizationDecks(orgId)) ?? {};
-          const orgInfo = await getOrganizationInfo(orgId);
+          const orgInfo = await getOrganizationPublicInfo(orgId);
           return Object.values(orgDecks).map(({ deck }) => ({
             ...deck,
             position: deck.position + positionOffset,
