+++ src/routes/api/playlists/user/+server.ts
@@ -1,4 +1,4 @@
-import { authenticate, getOrganizationInfo, getOrganizationPlaylists, getUserOrganizations, readPath } from '$lib/server/firebaseUtils';
+import { authenticate, getOrganizationPublicInfo, getOrganizationPlaylists, getUserOrganizationIds, readPath } from '$lib/server/firebaseUtils';
 import { json } from '@sveltejs/kit';
 import type { RequestHandler } from './$types';
 import { transformPlaylistWordsForClient } from '$lib/utils';
@@ -7,12 +7,12 @@ export const GET = (async (event) => {
   const { uid } = await authenticate(event);
   const userPlaylists = (await readPath<Database.Playlists.User>(`/playlists/user/${uid}`)) || {};
   const userPlaylistsArray = Object.values(userPlaylists).map((playlist) => playlist);
-  const organizationIds = await getUserOrganizations(uid);
+  const organizationIds = await getUserOrganizationIds(uid);
   const organizationPlaylistsArray = (
     await Promise.all(
       organizationIds.map(async (orgId) => {
         const orgPlaylists = (await getOrganizationPlaylists(orgId)) ?? {};
-        const orgInfo = await getOrganizationInfo(orgId);
+        const orgInfo = await getOrganizationPublicInfo(orgId);
         return Object.values(orgPlaylists).map((playlist) => ({
           ...playlist.playlist,
           orgSource: {
